muchos arreglos

+++ pedido/src/main/java/com/perfulandia/pedido/model/Detalle.java
@@ -23,7 +23,7 @@ public class Detalle {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idDetalle;
+    private Long idDetalle;
 
     private int idProducto; //Simulando la clase producto
