+++ pedido/src/main/java/com/perfulandia/pedido/repository/DetalleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.perfulandia.pedido.model.Detalle;
 
 @Repository
-public interface DetalleRepository extends JpaRepository<Detalle, Integer> {
+public interface DetalleRepository extends JpaRepository<Detalle, Long> {
     
 }
